HAWQ-269. Change net_disk_ratio from int to double type.

+++ src/include/utils/guc.h
@@ -391,7 +391,7 @@ extern double optimizer_cost_threshold;
 extern double optimizer_nestloop_factor;
 extern double locality_upper_bound;
 extern bool optimizer_cte_inlining;
-extern int net_disk_ratio;
+extern double net_disk_ratio;
 extern int optimizer_cte_inlining_bound;
 extern double optimizer_damping_factor_filter;
 extern double optimizer_damping_factor_join;
